* Migrate full target-building capabilities from transpile to generate_preset_pass_manager.
* Apply Matt's suggestions for custom mapping and control flow op names.

* Extend docstring, fix typos.
