+++ qiskit/transpiler/preset_passmanagers/__init__.py
@@ -1,6 +1,6 @@
 # This code is part of Qiskit.
 #
-# (C) Copyright IBM 2017, 2019.
+# (C) Copyright IBM 2017, 2024.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -21,7 +21,8 @@ This module contains functions for generating the preset pass managers
 for the transpiler. The preset pass managers are instances of
 :class:`~.StagedPassManager` which are used to execute the circuit
 transformations as part of Qiskit's compiler inside the
-:func:`~.transpile` function at the different optimization levels.
+:func:`~.transpile` function at the different optimization levels, but
+can also be used in a standalone manner.
 The functionality here is divided into two parts, the first includes the
 functions used generate the entire pass manager which is used by
 :func:`~.transpile` (:ref:`preset_pass_manager_generators`) and the
@@ -56,12 +57,20 @@ Stage Generator Functions
 .. autofunction:: generate_scheduling
 .. currentmodule:: qiskit.transpiler.preset_passmanagers
 """
+import copy
 
-import warnings
+from qiskit.circuit.controlflow import CONTROL_FLOW_OP_NAMES
+from qiskit.circuit.library.standard_gates import get_standard_gate_name_mapping
+from qiskit.providers.backend_compat import BackendV2Converter
+
+from qiskit.transpiler.instruction_durations import InstructionDurations
+from qiskit.transpiler.timing_constraints import TimingConstraints
 
 from qiskit.transpiler.passmanager_config import PassManagerConfig
-from qiskit.transpiler.target import target_to_backend_properties
+from qiskit.transpiler.target import Target, target_to_backend_properties
 from qiskit.transpiler import CouplingMap
+
+from qiskit.transpiler.exceptions import TranspilerError
 
 from .level0 import level_0_pass_manager
 from .level1 import level_1_pass_manager
@@ -91,16 +100,43 @@ def generate_preset_pass_manager(
     hls_config=None,
     init_method=None,
     optimization_method=None,
+    dt=None,
     *,
     _skip_target=False,
 ):
     """Generate a preset :class:`~.PassManager`
 
-    This function is used to quickly generate a preset pass manager. A preset pass
-    manager are the default pass managers used by the :func:`~.transpile`
+    This function is used to quickly generate a preset pass manager. Preset pass
+    managers are the default pass managers used by the :func:`~.transpile`
     function. This function provides a convenient and simple method to construct
     a standalone :class:`~.PassManager` object that mirrors what the transpile
+    function internally builds and uses.
 
+    The target constraints for the pass manager construction can be specified through a :class:`.Target`
+    instance, a `.BackendV1` or `.BackendV2` instance, or via loose constraints (``basis_gates``,
+    ``inst_map``, ``coupling_map``, ``backend_properties``, ``instruction_durations``,
+    ``dt`` or ``timing_constraints``).
+    The order of priorities for target constraints works as follows: if a ``target``
+    input is provided, it will take priority over any ``backend`` input or loose constraints.
+    If a ``backend`` is provided together with any loose constraint
+    from the list above, the loose constraint will take priority over the corresponding backend
+    constraint. This behavior is independent of whether the ``backend`` instance is of type
+    :class:`.BackendV1` or :class:`.BackendV2`, as summarized in the table below. The first column
+    in the table summarizes the potential user-provided constraints, and each cell shows whether
+    the priority is assigned to that specific constraint input or another input
+    (`target`/`backend(V1)`/`backend(V2)`).
+
+    ============================ ========= ======================== =======================
+    User Provided                target    backend(V1)              backend(V2)
+    ============================ ========= ======================== =======================
+    **basis_gates**              target    basis_gates              basis_gates
+    **coupling_map**             target    coupling_map             coupling_map
+    **instruction_durations**    target    instruction_durations    instruction_durations
+    **inst_map**                 target    inst_map                 inst_map
+    **dt**                       target    dt                       dt
+    **timing_constraints**       target    timing_constraints       timing_constraints
+    **backend_properties**       target    backend_properties       backend_properties
+    ============================ ========= ======================== =======================
 
     Args:
         optimization_level (int): The optimization level to generate a
@@ -126,16 +162,57 @@ def generate_preset_pass_manager(
             and ``backend_properties``.
         basis_gates (list): List of basis gate names to unroll to
             (e.g: ``['u1', 'u2', 'u3', 'cx']``).
-        inst_map (InstructionScheduleMap): Mapping object that maps gate to schedules.
+        inst_map (InstructionScheduleMap): Mapping object that maps gates to schedules.
             If any user defined calibration is found in the map and this is used in a
             circuit, transpiler attaches the custom gate definition to the circuit.
             This enables one to flexibly override the low-level instruction
             implementation.
         coupling_map (CouplingMap or list): Directed graph represented a coupling
-            map.
-        instruction_durations (InstructionDurations): Dictionary of duration
-            (in dt) for each instruction.
+            map. Multiple formats are supported:
+
+            #. ``CouplingMap`` instance
+            #. List, must be given as an adjacency matrix, where each entry
+               specifies all directed two-qubit interactions supported by backend,
+               e.g: ``[[0, 1], [0, 3], [1, 2], [1, 5], [2, 5], [4, 1], [5, 3]]``
+
+        instruction_durations (InstructionDurations or list): Dictionary of duration
+            (in dt) for each instruction. If specified, these durations overwrite the
+            gate lengths in ``backend.properties``. Applicable only if ``scheduling_method``
+            is specified.
+            The format of ``instruction_durations`` must be as follows:
+            They must be given as an :class:`.InstructionDurations` instance or a list of tuples
+
+            ```
+            [(instruction_name, qubits, duration, unit), ...].
+            | [('cx', [0, 1], 12.3, 'ns'), ('u3', [0], 4.56, 'ns')]
+            | [('cx', [0, 1], 1000), ('u3', [0], 300)]
+            ```
+
+            If ``unit`` is omitted, the default is ``'dt'``, which is a sample time depending on backend.
+            If the time unit is ``'dt'``, the duration must be an integer.
+        dt: Backend sample time (resolution) in seconds.
+            If provided, this value will overwrite the ``dt`` value in ``instruction_durations``.
+            If ``None`` (default) and a backend is provided, ``backend.dt`` is used.
         timing_constraints (TimingConstraints): Hardware time alignment restrictions.
+             A quantum computer backend may report a set of restrictions, namely:
+
+                - granularity: An integer value representing minimum pulse gate
+                  resolution in units of ``dt``. A user-defined pulse gate should have
+                  duration of a multiple of this granularity value.
+                - min_length: An integer value representing minimum pulse gate
+                  length in units of ``dt``. A user-defined pulse gate should be longer
+                  than this length.
+                - pulse_alignment: An integer value representing a time resolution of gate
+                  instruction starting time. Gate instruction should start at time which
+                  is a multiple of the alignment value.
+                - acquire_alignment: An integer value representing a time resolution of measure
+                  instruction starting time. Measure instruction should start at time which
+                  is a multiple of the alignment value.
+
+                This information will be provided by the backend configuration.
+                If the backend doesn't have any restriction on the instruction time allocation,
+                then ``timing_constraints`` is None and no adjustment will be performed.
+
         initial_layout (Layout | List[int]): Initial position of virtual qubits on
             physical qubits.
         layout_method (str): The :class:`~.Pass` to use for choosing initial qubit
@@ -205,8 +282,74 @@ def generate_preset_pass_manager(
         ValueError: if an invalid value for ``optimization_level`` is passed in.
     """
 
-    if coupling_map is not None and not isinstance(coupling_map, CouplingMap):
-        coupling_map = CouplingMap(coupling_map)
+    if backend is not None and getattr(backend, "version", 0) <= 1:
+        # This is a temporary conversion step to allow for a smoother transition
+        # to a fully target-based transpiler pipeline while maintaining the behavior
+        # of `transpile` with BackendV1 inputs.
+        backend = BackendV2Converter(backend)
+
+    # Check if a custom inst_map was specified before overwriting inst_map
+    _given_inst_map = bool(inst_map)
+    # If there are no loose constraints => use backend target if available
+    _no_loose_constraints = (
+        basis_gates is None
+        and coupling_map is None
+        and dt is None
+        and instruction_durations is None
+        and backend_properties is None
+        and timing_constraints is None
+    )
+    # If it's an edge case => do not build target
+    _skip_target = (
+        target is None
+        and backend is None
+        and (basis_gates is None or coupling_map is None or instruction_durations is not None)
+    )
+
+    # Resolve loose constraints case-by-case against backend constraints.
+    # The order of priority is loose constraints > backend.
+    dt = _parse_dt(dt, backend)
+    instruction_durations = _parse_instruction_durations(backend, instruction_durations, dt)
+    timing_constraints = _parse_timing_constraints(backend, timing_constraints)
+    inst_map = _parse_inst_map(inst_map, backend)
+    # The basis gates parser will set _skip_target to True if a custom basis gate is found
+    # (known edge case).
+    basis_gates, name_mapping, _skip_target = _parse_basis_gates(
+        basis_gates, backend, inst_map, _skip_target
+    )
+    coupling_map = _parse_coupling_map(coupling_map, backend)
+
+    if target is None:
+        if backend is not None and _no_loose_constraints:
+            # If a backend is specified without loose constraints, use its target directly.
+            target = backend.target
+        elif not _skip_target:
+            # Only parse backend properties when the target isn't skipped to
+            # preserve the former behavior of transpile.
+            backend_properties = _parse_backend_properties(backend_properties, backend)
+            # Build target from constraints.
+            target = Target.from_configuration(
+                basis_gates=basis_gates,
+                num_qubits=backend.num_qubits if backend is not None else None,
+                coupling_map=coupling_map,
+                # If the instruction map has custom gates, do not give as config, the information
+                # will be added to the target with update_from_instruction_schedule_map
+                inst_map=inst_map if inst_map and not inst_map.has_custom_gate() else None,
+                backend_properties=backend_properties,
+                instruction_durations=instruction_durations,
+                concurrent_measurements=(
+                    backend.target.concurrent_measurements if backend is not None else None
+                ),
+                dt=dt,
+                timing_constraints=timing_constraints,
+                custom_name_mapping=name_mapping,
+            )
+
+    # Update target with custom gate information. Note that this is an exception to the priority
+    # order (target > loose constraints), added to handle custom gates for scheduling passes.
+    if target is not None and _given_inst_map and inst_map.has_custom_gate():
+        target = copy.deepcopy(target)
+        target.update_from_instruction_schedule_map(inst_map)
 
     if target is not None:
         if coupling_map is None:
@@ -260,6 +403,119 @@ def generate_preset_pass_manager(
     else:
         raise ValueError(f"Invalid optimization level {optimization_level}")
     return pm
+
+
+def _parse_basis_gates(basis_gates, backend, inst_map, skip_target):
+    name_mapping = {}
+    standard_gates = get_standard_gate_name_mapping()
+    # Add control flow gates by default to basis set
+    default_gates = {"measure", "delay", "reset"}.union(CONTROL_FLOW_OP_NAMES)
+
+    try:
+        instructions = set(basis_gates)
+        for name in default_gates:
+            if name not in instructions:
+                instructions.add(name)
+    except TypeError:
+        instructions = None
+
+    if backend is None:
+        # Check for custom instructions
+        if instructions is None:
+            return None, name_mapping, skip_target
+
+        for inst in instructions:
+            if inst not in standard_gates or inst not in default_gates:
+                skip_target = True
+                break
+
+        return list(instructions), name_mapping, skip_target
+
+    instructions = instructions or backend.operation_names
+    name_mapping.update(
+        {name: backend.target.operation_from_name(name) for name in backend.operation_names}
+    )
+
+    # Check for custom instructions before removing calibrations
+    for inst in instructions:
+        if inst not in standard_gates or inst not in default_gates:
+            skip_target = True
+            break
+
+    # Remove calibrated instructions, as they will be added later from the instruction schedule map
+    if inst_map is not None and not skip_target:
+        for inst in inst_map.instructions:
+            for qubit in inst_map.qubits_with_instruction(inst):
+                entry = inst_map._get_calibration_entry(inst, qubit)
+                if entry.user_provided and inst in instructions:
+                    instructions.remove(inst)
+
+    return list(instructions) if instructions else None, name_mapping, skip_target
+
+
+def _parse_inst_map(inst_map, backend):
+    # try getting inst_map from user, else backend
+    if inst_map is None and backend is not None:
+        inst_map = backend.target.instruction_schedule_map()
+    return inst_map
+
+
+def _parse_backend_properties(backend_properties, backend):
+    # try getting backend_props from user, else backend
+    if backend_properties is None and backend is not None:
+        backend_properties = target_to_backend_properties(backend.target)
+    return backend_properties
+
+
+def _parse_dt(dt, backend):
+    # try getting dt from user, else backend
+    if dt is None and backend is not None:
+        dt = backend.target.dt
+    return dt
+
+
+def _parse_coupling_map(coupling_map, backend):
+    # try getting coupling_map from user, else backend
+    if coupling_map is None and backend is not None:
+        coupling_map = backend.coupling_map
+
+    # coupling_map could be None, or a list of lists, e.g. [[0, 1], [2, 1]]
+    if coupling_map is None or isinstance(coupling_map, CouplingMap):
+        return coupling_map
+    if isinstance(coupling_map, list) and all(
+        isinstance(i, list) and len(i) == 2 for i in coupling_map
+    ):
+        return CouplingMap(coupling_map)
+    else:
+        raise TranspilerError(
+            "Only a single input coupling map can be used with generate_preset_pass_manager()."
+        )
+
+
+def _parse_instruction_durations(backend, inst_durations, dt):
+    """Create a list of ``InstructionDuration``s. If ``inst_durations`` is provided,
+    the backend will be ignored, otherwise, the durations will be populated from the
+    backend.
+    """
+    final_durations = InstructionDurations()
+    if not inst_durations:
+        backend_durations = InstructionDurations()
+        if backend is not None:
+            backend_durations = backend.instruction_durations
+        final_durations.update(backend_durations, dt or backend_durations.dt)
+    else:
+        final_durations.update(inst_durations, dt or getattr(inst_durations, "dt", None))
+    return final_durations
+
+
+def _parse_timing_constraints(backend, timing_constraints):
+    if isinstance(timing_constraints, TimingConstraints):
+        return timing_constraints
+    if backend is None and timing_constraints is None:
+        timing_constraints = TimingConstraints()
+    elif backend is not None:
+        timing_constraints = backend.target.timing_constraints()
+    return timing_constraints
 
 
 __all__ = [
